Authorization consent token flow updated

+++ api/product/delete.php
@@ -6,28 +6,38 @@ header("Access-Control-Allow-Methods: DELETE");
 header("Content-Type: application/json; charset=UTF-8");
 
 require_once "../../config/Database.php";
+require_once "../../config/config.php";
 require_once "../../models/Product.php";
-require_once "../validate.php";
+require_once "../../models/Token.php";
+require_once "../../helper/helper.php";
 
 $db = new DbConnect();
 $conn = $db->connect();
 
 $product = new Product($conn);
 
-$prod = json_decode(file_get_contents("php://input"));
+//Valid the access token
+$response = validateToken($conn); 
+if($response['status'] === "Success"){
+    $prod = json_decode(file_get_contents("php://input"));
 
-try{
-    $res = $product->delete($prod->id);
+    try{
+        $res = $product->delete($prod->id);
 
-    $response['status'] = 'Success';
-    $response['message'] = 'Product deleted successfully';
+        $response['status'] = 'Success';
+        $response['message'] = 'Product deleted successfully';
 
-    echo json_encode($response);
-}catch(Exception $e){
-    $response['status'] = 'Failed';
-    $response['message'] = $e->getMessage();
+        echo json_encode($response);
+    }catch(Exception $e){
+        $response['status'] = 'Failed';
+        $response['message'] = $e->getMessage();
 
+        echo json_encode($response);
+    }
+}else{
     echo json_encode($response);
 }
+
+
